상품 서비스 리팩토링
